Updated Navbar and App for data to flow better. Updated Login flow and stored User state in cache
Navbar had pages as children cause routes were in it. moved routes to App and kept links in navbar, it sees routes in parent so still works and make it so main content is child of app not of nav bar which makes more sense.

Updated Login flow by lifting state up from App instead of navbar.